Migrate UI from WPF to Avalonia
This alternative to WPF is multi-platform.
This way the ReplayViewer can be used under Linux as well.

There was a problem with redrawing WriteableBitmaps. After the lock was disposed the image was not correctly invalidated. A part (top and left) of the image was not updated. A simple OnPropertyChanged did not work, since the underlying bitmap object remained the same. This is solved by using double buffering. Two bitmaps are kept and alternated every update.

Full screen mode of QuestDashboard is not correct yet.