Refactor MapGenerator
Do not recompute available positions every room, but use a set that is updated when a room is added.
This should allow higher performance, but tuning is still needed.
A benchmark project is added to check this.

Furthermore, walls are only added at the end, so no cleanup of inner walls is needed.