Fix double quest dashboard bug (#2)
* Change filter of dashboard to allow multiple quests.

Only one active quest was used to filter incoming game updates. Changed it to a set that is updated when games are added or removed.

* Do not allow users to queue a new quest when already active

Also refactored the check if a quest should be finished and removed. There is still a bug that sometimes it is possible 3 quests are active instead of the configured 2.

* Make thread-safe using locks

* Show result in game finished overlay

* Fix tests

When quest is finished, no new level is entered.

* Refactor creating game state on exceptions

* Small fixes after review

Instead of removing a specific game, a random game was removed. Changed bag to set that is manually locked, to fix this.

Also fixed the finished result not showing up at the end due to a missing PropertyChanged notification.