Rework sessions events (#11)
The game added and updated events can be replaced by the postman events that were already handled. The statistics are moved to the DashboardService, so GameServer becomes simpler.

Also the locking in GameServer is changed. Only the start of a quest is locked. The rest of the locks are replaced by using Concurrent collections.

For extra information an inner exception is added to the GameServerException, so when internal errors happen the original cause is logged.