Improve performance by reducing the use of Immutable collections (#12)
Create a new benchmark for game server performance. Parallel start is a bit wonky and takes a lot more time than expected. In particular if run together with the SingleGame test it takes about 8 seconds instead of < 1. Needs further investigation.

Made the generation of map stable of time. The ValueTuple hash code changes every restart of the application. By first sorting the entries in a collection before randomizing, makes it behave the same every time. Has only a slight performance penalty. A test is added that checks this. It is a single game run with fixed seed of 42 and then the python actions are logged and used.

Most optimizations have been achieved by reducing the usage of immutable collections in the MapGenerator class and using indexes instead of positions in locally used lists. Also the list of enemies in the Map class is converted (back) to separate properties to improve performance. Iterating over a list is slow vs accessing properties.