Remove more locks in GameServer and Game
Move the timeout check to the IsFinished flag, so the lock is no longer needed.
This means the status is no longer updated during the check, but state is only needed when actual used is communicating, which is with a call to Act. At the start of Act the timeout is checked and the status updated accordingly.