PickOne has changed so update tests accordingly
SingleGame is changed again.

TwoPlayerCombat needed some extra work.
Two players approach the enemy, but the enemy now moved (randomly) on the path of one player, so the attack had to be changed. This let to a bug appearing that needed to be fixed. It was not allowed for two players to attack the same enemy when the first already killed it. This caused an "inventory empty" exception because the game tried to place a boulder.

Now, characters remain present in the game even if they have a zero or negative health, until the end of the step. Then the clean up, that was already implemented, moves them out of the way and drops loot.