Replace Position value tuple with dedicated struct (#21)
* Replace Position tuple with dedicated struct

* Improve shortest path algorithm in map generator

* Improve Room class

* Small improvements

* Fix enum name changed in complete bot

* Update tests since hash code and comparison of Position have changed

Basically reverting back to the initial situation before the "stable over time" change was made. Since the Position is now a struct with our own method is behave consistently and the extra Order in the PickOne is no longer needed.

* Extra test for unknown action

* Cleanup

* Further improve performance of map generator by reducing use of immutable collections

* PickOne has changed so update tests accordingly

SingleGame is changed again.

TwoPlayerCombat needed some extra work.
Two players approach the enemy, but the enemy now moved (randomly) on the path of one player, so the attack had to be changed. This let to a bug appearing that needed to be fixed. It was not allowed for two players to attack the same enemy when the first already killed it. This caused an "inventory empty" exception because the game tried to place a boulder.

Now, characters remain present in the game even if they have a zero or negative health, until the end of the step. Then the clean up, that was already implemented, moves them out of the way and drops loot.

* Add missing Dispose