Map generator interface refactor (#33)
* Refactor MapGenerator interface

I didn't really like the static interface. It also messed up the coverage report.

But for thread-safety a new instance of the map generator should be used.
So, I moved out the implementation of the map generator to MapGeneratorImpl and use that in the MapGenerator that now implements a regular interface.
This preserves the thread-safety while making it easier to inject.

* Enable more warnings

* Cleanup

* Fix tests

* Fix startup of server