Fix internal error when player steps in closing door
When the player move is processed the door is closed in the LeaveCell function. As a result the EnterCell function tries to enter a closed door position. This throws a NotImplementedExcepton.

This problem is resolved by processing the closing (and opening) of doors after both players have moved. They will now first enter a cell that is still open before it is being closed on top of them.