Fix internal error when door closes on new player position (#41)
* Move plate next to door in crush test

* Create test to capture bug

When player who stands on a plate, steps of the plate into the position of the door that is closed, then an internal error is triggered. This should have been a kill of the player.

* Fix internal error when player steps in closing door

When the player move is processed the door is closed in the LeaveCell function. As a result the EnterCell function tries to enter a closed door position. This throws a NotImplementedExcepton.

This problem is resolved by processing the closing (and opening) of doors after both players have moved. They will now first enter a cell that is still open before it is being closed on top of them.

* Prevent closing and opening of the same door

Both players (one with key other with plate) can interact with the same door. This leads to a close and an open of the same door, which can cause an exception in the converter. A closed door tile cannot be converted to a closed tile.