Check level input for map generator range (#42)
When quest is finished, the user level is 23 to indicate it finished the quest. But, the map generator cannot create a map for that level, so it will throw an out-of-range exception.

Added an extra check to prevent this and changed the StartResult value to INVALID_LEVEL for a more generic failure.