Add training seed (#40)
* Add train seed to proto

* Fix rebuild of publish target

The zip file cannot be created if it already exists.

* Use same seed during quest

At start a random seed is selected for Quest games and returned in the start result. This seed is used for each level in the quest. So the user can reproduce a single level of a quest.

* Fix .NET version to 9.0.200

To prevent package errors.