Implement final mode (#63)
* Implement final mode

First of all, waiting for a queued quest is now implemented server side. So, the clients no longer have to sleep themselves and their game will start as soon as possible.

Furthermore, the behavior of a second quest start has changed. If a quest is already active, and another start is requested by the same user, then the previous quest is cancelled and the new request is placed at the end of the queue.

A new final mode has been added. The server can be started with `--final user_id1,user_id2,...,user_id4` to start the special final mode. Only the provided users are allowed to start a quest; training is not allowed. The server will wait with starting each quest until all selected users have requested a started. Then all games will start at the same time. A countdown is displayed on the server console before starting. All these quests will use the same seed.

Finally, the dashboard has been changed to show the quests in a 2 column grid instead of tabs. And the user control has been updated to scale the player surroundings to the available heigh, making it look better in a grid.

Documentation is updated to describe new queue behavior.

Tests have been updated accordingly, and new tests have been added for the final mode. A few extra tests regarding queue reporting have been added as well.

Example bots have been updated as well.

* Update Go bot to new qeueing behavior

* Revert changes to complete TS bot proto file

Updated ProtoStripper to allow specifying the level to strip. Complete TS bot is limited to level 20.

* Make a dedicated START_RESULT_NOT_ALLOWED

For use in final mode when a user tries to connect that is not in the list, or tries to start a training, or already started a quest.

* Allow injecting seed in final game server

* Add more status checks to GameServer tests

* Revert removal of QUEST_QUEUED documentation

* Fix missing newline at end of proto file