Move character parameters to Infra
They are used during creation of the Map, which is in Infra, and not by the Server directly. The ones in Parameters.cs weren't used anymore. So, it seems better to keep them in MutableCharacter and remove them from Paramters.

Also reduced nr of quests to 1, will increase when needed, but for now keep this restriction. Dashboard does not look right with two quests.

Fixes #73