Show when user has finished the quest
New fields added to the User database item, to store whether or not they finished the quest and which game was their best game.

When quest is finished by a user:
- in all the high scores (dashboard and portal) a star is shown besides their name,
- and an extra banner is shown on the user page.

For each improvement the game id of the quest with which that was achieved is stored in the database and also shown on the user page.

Fixes #68