Show overlay with error message in ReplayViewer
Instead of relying on Console.WriteLine, which does not work under Windows, an extra error handling interface has been added. When an ErrorMessage is set an overlay is shown with the error message.

The load button remains enabled, so the user can try again.

Also made sure that when nothing is loaded, nothing is shown. The progress bar and buttons are disabled. No longer showing "Game Finished".

Fixes #74