Small cleanup (#75)
* Small cleanup

- Removed todo.md file. All relevant actions have been added as GitHub issues.
- Removed/fixed some TODOs.
- Updated surroundings.png in Documentation to new tileset.

* Move character parameters to Infra

They are used during creation of the Map, which is in Infra, and not by the Server directly. The ones in Parameters.cs weren't used anymore. So, it seems better to keep them in MutableCharacter and remove them from Paramters.

Also reduced nr of quests to 1, will increase when needed, but for now keep this restriction. Dashboard does not look right with two quests.

Fixes #73

* Compute tile size based on bitmap size

Fixes #70

* Add C++ bot to update_protos script

* Remove mentions to env.json

* Show "training completed" when only a single level is finished.

A sequence of levels (i.e. initial level of replay is not the same as the end level) will only happen during a quest.
If a quest finishes the "quest finished" message will be shown.

Fixes #76

* Show when user has finished the quest

New fields added to the User database item, to store whether or not they finished the quest and which game was their best game.

When quest is finished by a user:
- in all the high scores (dashboard and portal) a star is shown besides their name,
- and an extra banner is shown on the user page.

For each improvement the game id of the quest with which that was achieved is stored in the database and also shown on the user page.

Fixes #68