Prevent Use actions in first levels (#77)
Up till level 2 the proto file does not mention the Use actions.
These are the only levels that do not contain anything to use on (enemies, boulders or closed doors), so a simple check is added at the start of the game whether or not use is allowed.
A test is added to cover this situation.

Fixes #72