Properly show canceled quests
When a quest is canceled the new game status is not sent to the dashboard, because no further interaction takes place at the GameService level. The dashboard does not get a new ActRequest/Response update, so it does not know the new game status. A new update event is added to the dashboard service to notify status changes. The game observation is adjusted accordingly. It is now only triggered when the a game is canceled, but should also be triggeren when a game times out. Will fix that at a later time.