Fix placement of exit (#99)
For random seed 899782328 in level 14 the exit was not properly placed inside a wall. There was a corridor crossing the bottom right corner of the exit room. This made the exit "float" in the middle, making the door useless.

Added an extra step to move the exit position inside the wall.

Also did some slight refactoring to always use the same placement function.