More improvements (#102)
* Use standard logger for final round info

This way the command line output is hidden during tests.

* Store quest progress in database

Every time a level is finished during a quest it is reported to the database.
On the portal start page the latest 20 quest results are shown.

Also changed the Id field of each data type to a non-nullable type. This reduces some annoying checks.

* Show quest history on user page

* Provide download links to quest .swoq files

* Improve portal with logo, header and other colors.

* Override ReplayStorage setting from environment variable

* Prefix outputs with Swoq.

* Fix unit tests

* Fix stylesheet filename

* Make sure concurrent read/write is allowed on replay files

* Prevent downloading recently updated replays

They might be still in use causing exceptions during the read.

* Make download links not change the URL of the browser

This way you can refresh and the page will reload instead of redownloading the file.

* Fix download links

Use the `download` attribute on the links. This way clicking the link won't change the browser address.

Furthermore, by reading the file content manually the file can be read while in progress.

* Use more of MongoDB features in getting quest history