Small performance improvements and last updates (#103)
* Update packages to latest versions

* Replace DateTime.Now with UtcNow for tiny performance improvement

* Stop FastBot when connection drops

* Improve performance of ReplaySaver by keeping file open

As long as the output file is the same (with only one Quest active that is most of the time), the file stream is left open.

* Apply some performance optimizations

- Allocate player surroundings once and update
- Change Cell to Tile mapping to a dictionary for faster lookup
- Cache player positions

* Remove Bold from textblocks with emojis to preserve color

By having the bold flag, the emojis were changed to flat white images instead of the colored ones as before.

* Fix command-line loading of ReplayViewer

ViewModels must be created and assigned to properties on the UI thread.

* Add readme to ReplayViewer

* Update main readme