Fixes after weekend (#105)
* Fix SingleGame benchmark

Level is now added to seed, so the used seed was wrong.
Cleanup / Dispose is not called by BenchmarkDotNet automatically, you have to use the attributes. You should actually also use [GlobalSetup], but that will give issues with nullable checks. The Cleanup is only called once, at the end, so it is effectively the same.

* Downloads replays in high scores list

* Remove some local information